requirements.txt
 python-docx
 spacy
 presidio-analyzer
 langchain
 chromadb
 ollama-python
 sentence-transformers
chromadb
sentence-transformers
